Tests/LibWeb/Text/expected/css/content-visibility-interpolation.txt

progress 0: visible
progress 0.4: visible
progress 0.6: auto
progress 1: auto
